Fixed issue
Fixed issue: Users with Optifine/Modloader/Reis minmap/MorePlayerModels
have the issue that if they shift right click on crafting result the
client fires for each item in the workbench grid a craft event. So the
plugin was confused ;) But this is now fixed by correcting the
calculation of result times.